Event CRUD classes made.

+++ Semester 3 project/src/main/java/sem3/project/individual/business/DeleteAccountsFunctionality.java
@@ -2,5 +2,5 @@ package sem3.project.individual.business;
 
 public interface DeleteAccountsFunctionality
 {
-    void deleteAccount(int id);
+    void deleteAccount(Long id);
 }
