+++ Semester 3 project/src/main/java/sem3/project/individual/domain/accounts/UpdateAccountRequest.java
@@ -6,7 +6,7 @@ import lombok.Getter;
 @Getter @AllArgsConstructor
 public class UpdateAccountRequest
 {
-    private final int targetId;
+    private final Long targetId;
     private String username;
     private String password;
     private String email;
